+++ devices/webcam/pose_vis/extensions/hands.py
@@ -12,7 +12,7 @@ from mediapipe.framework.formats.landmark_pb2 import NormalizedLandmarkList
 
 # Every extension will probably need these imports
 from pose_vis.extension import PoseVisExtension, ExtensionResult
-from pose_vis.video_stream import StreamMetaData
+from pose_vis.streams.messages import StreamMetaData
 from argparse import ArgumentParser, Namespace
 
 from typing import Optional, Tuple
