+++ devices/webcam/pose_vis/streams/camera_stream.py
@@ -6,9 +6,9 @@ import cv2
 import labgraph as lg
 import numpy as np
 
-from pose_vis.video_stream import ProcessedVideoFrame, StreamMetaData
+from pose_vis.streams.messages import ProcessedVideoFrame, StreamMetaData, CombinedExtensionResult
 from pose_vis.frame_processor import FrameProcessor
-from pose_vis.extension import PoseVisExtension, CombinedExtensionResult
+from pose_vis.extension import PoseVisExtension
 from pose_vis.performance_utility import PerfUtility
 from typing import Optional, Tuple, List
