+++ devices/webcam/pose_vis/streams/replay_stream.py
@@ -4,9 +4,9 @@
 import asyncio
 import labgraph as lg
 
-from pose_vis.video_stream import ProcessedVideoFrame, StreamMetaData
+from pose_vis.streams.messages import ProcessedVideoFrame, StreamMetaData, CombinedExtensionResult
 from pose_vis.frame_processor import FrameProcessor
-from pose_vis.extension import PoseVisExtension, CombinedExtensionResult
+from pose_vis.extension import PoseVisExtension
 from pose_vis.performance_utility import PerfUtility
 from typing import Optional, List
 from labgraph.loggers.hdf5.reader import HDF5Reader
